check if file exists

+++ var/spack/repos/builtin/packages/sim-model/package.py
@@ -114,9 +114,11 @@ class SimModel(Package):
         nrnivmodl_params = self._nrnivmodlcore_params(include_flag, link_flag)
         with working_dir('build_' + self.mech_name, create=True):
             force_symlink(mods_location, 'mod')
-            os.rename(mods_location + "/adex.mod", mods_location + "/adex.mod.txt")
+            if os.path.exists(mods_location + "/adex.mod"):
+                os.rename(mods_location + "/adex.mod", mods_location + "/adex.mod.txt")
             which('nrnivmodl-core')(*(nrnivmodl_params + ['mod']))
-            os.rename(mods_location + "/adex.mod.txt", mods_location + "/adex.mod")
+            if os.path.exists(mods_location + "/adex.mod.txt"):
+                os.rename(mods_location + "/adex.mod.txt", mods_location + "/adex.mod")
             output_dir = os.path.basename(self.nrnivmodl_outdir)
             mechlib = find_libraries('libcorenrnmech' + self.lib_suffix + '*',
                                      output_dir)
@@ -155,9 +157,11 @@ class SimModel(Package):
                     which('nrnivmech_install.sh', path=".")(prefix)
                 else:
                     # Set dest to install
-                    os.rename(os.getcwd()+"/mod/adex.mod", os.getcwd()+"/mod/adex.mod.txt")
+                    if os.path.exists(os.getcwd()+"/mod/adex.mod"):
+                        os.rename(os.getcwd()+"/mod/adex.mod", os.getcwd()+"/mod/adex.mod.txt")
                     which('nrnivmodl-core')("-d", prefix, 'mod')
-                    os.rename(os.getcwd()+"/mod/adex.mod.txt", os.getcwd()+"/mod/adex.mod")
+                    if os.path.exists(os.getcwd()+"/mod/adex.mod.txt"):
+                        os.rename(os.getcwd()+"/mod/adex.mod.txt", os.getcwd()+"/mod/adex.mod")
 
         # Install special
         shutil.copy(join_path(arch, 'special'), prefix.bin)
